Finished updates to support the new MAU 1.1 firmware:
1) Adjusted Flight Profiles to match transition between sampling rates.
2) Created folder MainDev for the GraphGenerator and updated the program 
to support timestamps generation.

3) Updated the Excel sheet to support timestamps and adaptive chart 
lenght.

4) Moved old GraphGenerator and Excel sheet implementations to folder 
"old".

5) Added .bat files for reading and erasing the EEPROM to source 
control.
