+++ Firmware/MainDev/FlightProfile_500.h
@@ -100,35 +100,35 @@ const uint16_t pressure_samples[] PROGMEM = {
   26032,
   26019,
   26006,
-  25915,
-  25850,
-  25785,
-  25720,
-  25655,
-  25590,
-  25525,
-  25460,
-  25395,
-  25330,
-  25265,
-  25200,
-  25135,
-  25070,
-  25005,
-  24940,
-  24875,
-  24810,
-  24745,
-  24680,
-  24615,
-  24550,
-  24485,
-  24420,
-  24355,
-  24290,
-  24225,
-  24160,
-  24100,
+  25930,
+  25865,
+  25800,
+  25735,
+  25670,
+  25605,
+  25540,
+  25475,
+  25410,
+  25345,
+  25280,
+  25215,
+  25150,
+  25085,
+  25020,
+  24955,
+  24890,
+  24825,
+  24760,
+  24695,
+  24630,
+  24565,
+  24500,
+  24435,
+  24370,
+  24305,
+  24240,
+  24175,
+  24115,
   24099,
   24102,
   24099,
