+++ Firmware/MainDev/FlightProfile_Urano.h
@@ -141,10 +141,10 @@ const uint16_t pressure_samples[] PROGMEM = {
   28414,
   28410,
   28405,
-  28395,
-  28389,
-  28378,
-  28371,
+  28375, //28395,
+  //28389,
+  //28378,
+  //28371,
   28334,
   28286,
   28255,
@@ -251,5 +251,5 @@ const uint16_t pressure_samples[] PROGMEM = {
   25100,
   25101,
   25100
-  
+
 };
