--- GraphGenerator/Altitude Graph Generator.layout
@@ -1,8 +1,0 @@
-[Editors]
-Order=0
-Focused=0
-[Editor_0]
-CursorCol=53
-CursorRow=171
-TopLine=126
-LeftChar=1
