--- GraphGenerator/GraphGenerator_User.layout
@@ -1,8 +1,0 @@
-[Editors]
-Order=0
-Focused=0
-[Editor_0]
-CursorCol=1
-CursorRow=128
-TopLine=110
-LeftChar=1
