+++ GraphGenerator/MainDev/build.bat
@@ -1,0 +1,2 @@
+C:\MinGW\bin\gcc D:\Engineering\GitHub\MAU-1.0\GraphGenerator\MainDev\main.c -o C:\Leitor_MAU\GraphGenerator.exe
+timeout /t -1
