--- GraphGenerator/TestFilesRenamer.layout
@@ -1,8 +1,0 @@
-[Editors]
-Order=0
-Focused=0
-[Editor_0]
-CursorCol=4
-CursorRow=61
-TopLine=39
-LeftChar=1
