--- GraphGenerator/Zeroed_Memory_formatted.txt
@@ -1,1 +1,0 @@
-0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x90, 0x53, 0x9e, 0x7d, 0x90, 0x53, 0x90, 0x53, 0x5a, 0x6a, 0x1b, 0x68, 0x18, 0xfc, 0x78, 0x94, 0x98, 0xd6, 0xd0, 0xb, 0xea, 0x15, 0x7c, 0x0, 0xf9, 0xff, 0x8c, 0x3c, 0xf8, 0xc6, 0x70, 0x17
