--- GraphGenerator/flight_data_raw.txt
@@ -1,30 +1,0 @@
-27028
-27018
-27168
-28747
-29224
-29469
-29416
-29134
-28650
-27392
-27023
-27021
-27022
-27020
-27022
-27023
-27025
-27025
-27024
-27026
-27024
-27024
-27028
-27023
-27027
-27029
-27026
-27027
-27024
-27029
