--- GraphGenerator/flight_data_true.txt
@@ -1,36 +1,0 @@
-
-90440,66
-83868,70
-
-
-Pressao
-90440,66
-90467,62
-90063,33
-85811,00
-84527,65
-83868,70
-84011,24
-84769,75
-86072,05
-89459,70
-90454,14
-90459,53
-90456,83
-90462,22
-90456,83
-90454,14
-90448,75
-90448,75
-90451,44
-90446,05
-90451,44
-90451,44
-90440,66
-90454,14
-90443,36
-90437,97
-90446,05
-90443,36
-90451,44
-90437,97
