+++ GraphGenerator/old/GraphGenerator_User.layout
@@ -1,0 +1,10 @@
+[Editors]
+Order=0
+Focused=0
+[Editor_0]
+CursorCol=22
+CursorRow=3
+TopLine=1
+LeftChar=1
+Open=1
+Top=1
