+++ GraphGenerator/old/Makefile.win
@@ -1,0 +1,30 @@
+# Project: GraphGenerator_User
+# Makefile created by Dev-C++ 4.9.9.2
+
+CPP  = g++.exe -D__DEBUG__
+CC   = gcc.exe -D__DEBUG__
+WINDRES = windres.exe
+RES  = 
+OBJ  = GraphGenerator_User.o $(RES)
+LINKOBJ  = GraphGenerator_User.o $(RES)
+LIBS =  -L"C:/Dev-Cpp/lib"  -g3 
+INCS =  -I"C:/Dev-Cpp/include" 
+CXXINCS =  -I"C:/Dev-Cpp/lib/gcc/mingw32/3.4.2/include"  -I"C:/Dev-Cpp/include/c++/3.4.2/backward"  -I"C:/Dev-Cpp/include/c++/3.4.2/mingw32"  -I"C:/Dev-Cpp/include/c++/3.4.2"  -I"C:/Dev-Cpp/include" 
+BIN  = GraphGenerator_User.exe
+CXXFLAGS = $(CXXINCS)   -ansi -traditional-cpp -g3
+CFLAGS = $(INCS)   -ansi -traditional-cpp -g3
+RM = rm -f
+
+.PHONY: all all-before all-after clean clean-custom
+
+all: all-before GraphGenerator_User.exe all-after
+
+
+clean: clean-custom
+	${RM} $(OBJ) $(BIN)
+
+$(BIN): $(OBJ)
+	$(CC) $(LINKOBJ) -o "GraphGenerator_User.exe" $(LIBS)
+
+GraphGenerator_User.o: GraphGenerator_User.c
+	$(CC) -c GraphGenerator_User.c -o GraphGenerator_User.o $(CFLAGS)
